+++ src/main/java/fr/army/stelyteam/conversations/ConvWithdrawMoney.java
@@ -14,15 +14,15 @@ public class ConvWithdrawMoney extends StringPrompt {
     public Prompt acceptInput(ConversationContext con, String answer) {
         Player author = (Player) con.getForWhom();
         String teamID = StelyTeamPlugin.sqlManager.getTeamIDFromPlayer(author.getName());
-        Integer teamMoney = StelyTeamPlugin.sqlManager.getTeamMoney(teamID);
+        Float teamMoney = StelyTeamPlugin.sqlManager.getTeamMoney(teamID);
         EconomyManager eco = new EconomyManager();
-        Integer money;
+        Float money;
 
         if (answer.equals("all")){
             money = teamMoney;
         }else{
             try {
-                money = Integer.parseInt(answer);
+                money = Float.parseFloat(answer);
             } catch (NumberFormatException e) {
                 author.sendRawMessage("Veuillez entrer un nombre");
                 return null;
@@ -49,7 +49,7 @@ public class ConvWithdrawMoney extends StringPrompt {
     }
 
 
-    private boolean teamReachedMinMoney(String teamID, Integer money, Integer teamMoney) {
+    private boolean teamReachedMinMoney(String teamID, Float money, Float teamMoney) {
         return teamMoney - money < 0;
     }
 }
