+++ src/main/java/fr/army/stelyteam/utils/EconomyManager.java
@@ -10,16 +10,16 @@ import net.milkbowl.vault.economy.Economy;
 public class EconomyManager {
     public static Economy economy = null;
 
-    public boolean checkMoneyPlayer(Player player, int money) {
+    public boolean checkMoneyPlayer(Player player, float money) {
         return economy.getBalance(player) >= ((double) money);
     }
 
-    public void removeMoneyPlayer(Player player, int money) {
+    public void removeMoneyPlayer(Player player, float money) {
         economy.withdrawPlayer(player, money);
         player.sendMessage("Vous avez payé " + money + "€");
     }
 
-    public void addMoneyPlayer(Player player, int money) {
+    public void addMoneyPlayer(Player player, float money) {
         economy.depositPlayer(player, money);
         player.sendMessage("Vous avez reçu " + money + "€");
     }
